Contact components which are located in two layouts in client layer have been implemented completely.

+++ MyBlog/src/app/components/admin/admin-post/admin-post-form/admin-post-form.component.ts
@@ -140,10 +140,18 @@ export class AdminPostFormComponent implements OnInit {
   }
 
   onDeleteTag(tag: any): void {
-    this.tagService.deleteResource(tag.id).subscribe(response => {
-      const index = this.tags.indexOf(tag);
-      this.tags.splice(index, 1);
+    this.tagService.getTagUsage(tag.id).subscribe(response => {
+      if (response === true) {
+        return;
+      }
+      else {
+        this.tagService.deleteResource(tag.id).subscribe(response => {
+          const index = this.tags.indexOf(tag);
+          this.tags.splice(index, 1);
+        });
+      }
     });
+
   }
 
 }
